requirements.txt
Django==3.2.16
psycopg2-binary
python-dotenv
djangorestframework
django-cors-headers
whitenoise
django-storages[azure]
Pillow
